Split client and server builds
