+++ vitest.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from "vitest/config";
+import { pathAliases } from "./build.utils";
+
+export default defineConfig({
+  resolve: { alias: pathAliases },
+  test: { environment: "jsdom" },
+});
